@@ -2,6 +2,7 @@ package pubsub
 
 import (
 	"encoding/json"
+	"go.uber.org/zap"
 	"io"
 	"sync"
 	"time"
@@ -12,7 +13,7 @@ type PubConn interface {
 	SetWriteDeadline(time.Time) error
 }
 
-type TopicMap map[interface{}]struct{}
+type TopicMap map[any]struct{}
 
 type PubSub struct {
 	subscribers  sync.Map
@@ -62,9 +63,9 @@ func (p *PubSub) EvictAndClose(conn PubConn) {
 }
 
 // Publish data to the connections that subscribed to the subKey. If subKey is nil, it will be sent to all connections.
-func (p *PubSub) Publish(data interface{}, subKey interface{}) (err error) {
+func (p *PubSub) Publish(data any, subKey any) (err error) {
 	var mb []byte
-	p.subscribers.Range(func(conn, topic interface{}) bool {
+	p.subscribers.Range(func(conn, topic any) bool {
 		if subKey != nil {
 			if s := topic.(TopicMap); s != nil { //make sure topic not empty
 				if _, ok := s[subKey]; !ok {
@@ -89,4 +90,72 @@ func (p *PubSub) Publish(data interface{}, subKey interface{}) (err error) {
 		return true
 	})
 	return err
+}
+
+type delayPubEntry struct {
+	subKey any
+	data   any
+	t      time.Time
+}
+
+type DelayPublish struct {
+	mu sync.Mutex
+	*PubSub
+	delayPubEntries []delayPubEntry
+	blockInEmptyCh  chan struct{}
+	delay           time.Duration
+	logger          *zap.Logger
+}
+
+func NewDelayPublish(ps *PubSub, delay time.Duration, logger *zap.Logger) *DelayPublish {
+	p := &DelayPublish{
+		PubSub:         ps,
+		blockInEmptyCh: make(chan struct{}),
+		delay:          delay,
+		logger:         logger,
+	}
+	go p.run()
+	return p
+}
+
+func (p *DelayPublish) DelayPublish(data any, subKey any) {
+	if p.delay == 0 {
+		err := p.Publish(data, subKey)
+		if err != nil {
+			p.logger.WithOptions(zap.AddCallerSkip(1)).DPanic("publish", zap.Error(err))
+		}
+		return
+	}
+	p.mu.Lock()
+	defer p.mu.Unlock()
+	if len(p.delayPubEntries) == 0 {
+		p.blockInEmptyCh <- struct{}{}
+	}
+	p.delayPubEntries = append(p.delayPubEntries, delayPubEntry{subKey: subKey, data: data, t: time.Now().Add(p.delay)})
+}
+
+func (p *DelayPublish) run() {
+	var timer = time.NewTimer(100000 * time.Hour)
+	var err error
+	for {
+		if len(p.delayPubEntries) == 0 {
+			timer.Reset(100000 * time.Hour)
+		} else {
+			timer.Reset(p.delayPubEntries[0].t.Sub(time.Now()))
+		}
+		select {
+		case <-timer.C:
+			err = p.Publish(p.delayPubEntries[0].subKey, p.delayPubEntries[0].data)
+			if err != nil {
+				p.logger.DPanic("publish", zap.Error(err))
+			}
+			p.mu.Lock()
+			p.delayPubEntries = p.delayPubEntries[1:]
+			p.mu.Unlock()
+		case <-p.blockInEmptyCh:
+			if !timer.Stop() {
+				<-timer.C
+			}
+		}
+	}
 }
