@@ -129,7 +129,7 @@ func TestRemoveAllAvailable(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			tt.wantErr(t, RemoveAllAvailable(tt.args.upstreamId), fmt.Sprintf("RemoveAllAvailable(%v)", tt.args.upstreamId))
+			tt.wantErr(t, RemoveAvailableByUpstreamId(tt.args.upstreamId), fmt.Sprintf("RemoveAvailableByUpstreamId(%v)", tt.args.upstreamId))
 		})
 	}
 }
