@@ -3,17 +3,14 @@ package db
 import "github.com/google/uuid"
 
 type Upstream struct {
-	Id             int    `json:"id"`
-	Username       string `json:"username"`
-	Password       string `json:"password"`
-	Sync           string `json:"sync"`
-	Login          string `json:"login"`
-	DataHistory    string `json:"data_history"`
-	LatestSnapshot string `json:"latest_snapshot"`
+	Id       int    `json:"id"`
+	Username string `json:"username"`
+	Password string `json:"password"`
+	Url      string `json:"url"`
 }
 
 func GetUpstreams() ([]Upstream, error) {
-	rows, err := TideDB.Query(`select id, username, password, sync, login, data_history, latest_snapshot from upstreams`)
+	rows, err := TideDB.Query(`select id, username, password, url from upstreams`)
 	if err != nil {
 		return nil, err
 	}
@@ -22,7 +19,7 @@ func GetUpstreams() ([]Upstream, error) {
 		us []Upstream
 	)
 	for rows.Next() {
-		err = rows.Scan(&u.Id, &u.Username, &u.Password, &u.Sync, &u.Login, &u.DataHistory, &u.LatestSnapshot)
+		err = rows.Scan(&u.Id, &u.Username, &u.Password, &u.Url)
 		if err != nil {
 			return nil, err
 		}
@@ -36,11 +33,11 @@ func GetUpstreams() ([]Upstream, error) {
 func EditUpstream(up *Upstream) (err error) {
 	if up.Id == 0 {
 		err = TideDB.QueryRow(
-			`insert into upstreams (username, password, sync, login, data_history, latest_snapshot) values ($1,$2,$3,$4,$5,$6) returning id`,
-			up.Username, up.Password, up.Sync, up.Login, up.DataHistory, up.LatestSnapshot).Scan(&up.Id)
+			`insert into upstreams (username, password, url) values ($1,$2,$3) returning id`,
+			up.Username, up.Password, up.Url).Scan(&up.Id)
 	} else {
-		_, err = TideDB.Exec(`update upstreams set username=$1,password=$2,sync=$3,login=$4,data_history=$5 where id=$6`,
-			up.Username, up.Password, up.Sync, up.Login, up.DataHistory, up.Id)
+		_, err = TideDB.Exec(`update upstreams set username=$1, password=$2, url=$3 where id=$4`,
+			up.Username, up.Password, up.Url, up.Id)
 	}
 	return err
 }
@@ -88,7 +85,7 @@ func DelUpstream(id int) ([]uuid.UUID, error) {
 
 func GetStationsByUpstreamId(upId int) ([]Station, error) {
 	rows, err := TideDB.Query(`select id, identifier, name, ip_addr, location, partner, cameras, status, status_changed_at from stations 
-inner join upstream_stations on stations.id = upstream_stations.station_id where stations.deleted_at is null and upstream_stations.upstream_id=$1`, upId)
+inner join upstream_stations on stations.id = upstream_stations.station_id where upstream_stations.upstream_id=$1 and stations.deleted_at is null`, upId)
 	if err != nil {
 		return nil, err
 	}
@@ -115,7 +112,7 @@ func IsUpstreamStation(stationId uuid.UUID) (upstream bool, err error) {
 }
 
 func GetUpstreamsByStationId(stationId uuid.UUID) ([]Upstream, error) {
-	rows, err := TideDB.Query(`select id, username, password, login, sync, data_history, latest_snapshot from upstream_stations join upstreams on upstream_stations.upstream_id = upstreams.id where station_id=$1`, stationId)
+	rows, err := TideDB.Query(`select id, username, password, url from upstream_stations join upstreams on upstream_stations.upstream_id = upstreams.id where station_id=$1`, stationId)
 	if err != nil {
 		return nil, err
 	}
@@ -124,7 +121,7 @@ func GetUpstreamsByStationId(stationId uuid.UUID) ([]Upstream, error) {
 		us []Upstream
 	)
 	for rows.Next() {
-		err = rows.Scan(&u.Id, &u.Username, &u.Password, &u.Login, &u.Sync, &u.DataHistory, &u.LatestSnapshot)
+		err = rows.Scan(&u.Id, &u.Username, &u.Password, &u.Url)
 		if err != nil {
 			return nil, err
 		}
